requirements.txt
asgiref==3.8.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.2.23
djangorestframework==3.16.0
idna==3.4
oauthlib==3.2.2
Pillow==9.2.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.6.0
python-decouple==3.8
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
typing_extensions==4.14.0
urllib3==1.26.12
